Refactor about
