+++ src/Components/about/aboutContainer.tsx
@@ -1,0 +1,9 @@
+export const day: number = 100;
+export const cnt_comment: number = 120;
+
+export const heading: string[] = [
+  `${day}일 동안`,
+  `${cnt_comment}개의 불편함이`,
+  `모였습니다.`,
+];
+export const explanation: string[] = ["학교가 불편한 순간"];
