+++ src/Components/about/styled.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const AboutBox = styled.div`
+  margin: 230px 30px 0 9.5%;
+  width: 100%;
+  height: 100%;
+`;
